% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NRSF.anno}
\alias{NRSF.anno}
\title{Annotations of 4,828 NRSF (REST) motif instances.}
\format{A data frame with 4828 rows and 6 variables:
\describe{
  \item{chrom}{chromosome}
  \item{start}{starting base pair (1-based, inclusive)}
  \item{end}{ending base pair}
  \item{strand}{strand (\code{"-"} or \code{"+"})}
  \item{score}{Position Weight Matrix score (log-likelihood)}
  \item{signalValue}{ChIP-seq signal value in K562 cells}
}}
\source{
Motif instances are downloaded from \url{http://homer.salk.edu/homer/} (HOMER Known Motifs track).

ChIP-seq signal value was extracted from \url{http://hgdownload.cse.ucsc.edu/goldenPath/hg19/encodeDCC/wgEncodeAwgTfbsUniform/wgEncodeAwgTfbsHaibK562NrsfV0416102UniPk.narrowPeak.gz}.
}
\usage{
NRSF.anno
}
\description{
A dataset containing the annotations of 4,828 NRSF (REST) motif instances in the human genome (hg19 assembly).
}
\keyword{datasets}
