% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NRSF.margin}
\alias{NRSF.margin}
\title{Number of base pairs of margin for \code{NRSF.cuts}.}
\format{Integer, equal to 200 (base pairs).}
\usage{
NRSF.margin
}
\description{
Number of base pairs of upstream and downstream margin for \code{NRSF.cuts}.
}
\keyword{datasets}

